+++ buildSrc/src/main/groovy/ca/allanwang/kau/Versions.groovy
@@ -26,11 +26,11 @@ class Versions {
     // https://kotlinlang.org/docs/reference/using-gradle.html
     static def kotlin = '1.3.11'
 
+    // https://github.com/Kotlin/kotlinx.coroutines/releases
+    static def coroutines = '1.0.1'
+
     // https://github.com/mikepenz/AboutLibraries/releases
     static def aboutLibraries = '6.2.0'
-
-    // https://github.com/Kotlin/anko/releases
-    static def anko = '0.10.5'
 
     // https://github.com/wasabeef/Blurry/releases
     static def blurry = '2.1.1'